requirements.txt
streamlit
streamlit-lottie
python-dotenv
pandas
openpyxl
xlrd>=2.0.1
